feat: Simplication
Now compiles, the bots original code used the anyhow crate and it couldn't get
the main() return type correct. I searched for "tokio hello world app" and found
https://github.com/tokio-rs/tokio/blob/master/examples/hello_world.rs.

It suggested the `Result<(), Box<dyn Error>>` as the result, I copied the hello_world.rs
code in to the bot and it resulted in this commit and it now compiles!

the `fn main` needs refactoring but a future change :)